movie slice1,2 수정 완료
