+++ src/store/modules/movieSlice2.js
@@ -1,87 +1,55 @@
 import { createSlice } from "@reduxjs/toolkit";
-import { getContent, getFilteredContent } from "./getThunk2";
+import { getDirectorsByCountry } from "./getThunk3";
 
 const initialState = {
-    data: [],
-    currentPage: 1,
-    totalPages: 1,
+    filteredDirectors: [],
     loading: false,
     error: null,
-    activeFilters: {
-        sortBy: "popularity.desc",
-        genres: [],
-        year: null,
-        country: null,
-        runtime: null,
-        ratings: null,
-    },
+    hasMore: true,
+    currentPage: 1,
+    selectedCountry: "all", // 국가 선택 상태 유지
 };
 
-const movieSlice = createSlice({
-    name: "movie",
+const directorSlice = createSlice({
+    name: "director",
     initialState,
     reducers: {
-        setPage: (state, action) => {
-            state.currentPage = action.payload;
-        },
-        clearData: (state) => {
-            state.data = [];
+        setSelectedCountry: (state, action) => {
+            state.selectedCountry = action.payload;
+            state.filteredDirectors = [];
             state.currentPage = 1;
-            state.totalPages = 1;
-        },
-        setFilters: (state, action) => {
-            state.activeFilters = action.payload;
-        },
-        clearFilters: (state) => {
-            state.activeFilters = initialState.activeFilters;
+            state.hasMore = true;
         },
     },
     extraReducers: (builder) => {
         builder
-            .addCase(getContent.pending, (state) => {
+            .addCase(getDirectorsByCountry.pending, (state) => {
                 state.loading = true;
                 state.error = null;
             })
-            .addCase(getContent.fulfilled, (state, action) => {
+            .addCase(getDirectorsByCountry.fulfilled, (state, action) => {
+                // 첫 페이지일 경우 데이터 초기화, 아닐 경우 기존 데이터에 추가
                 if (action.payload.currentPage === 1) {
-                    state.data = action.payload.data;
+                    state.filteredDirectors = action.payload.filteredDirectors;
                 } else {
-                    const newData = action.payload.data.filter(
-                        (newItem) => !state.data.some((existingItem) => existingItem.id === newItem.id)
+                    // 중복 제거를 위해 id 기준으로 필터링
+                    const newDirectors = action.payload.filteredDirectors.filter(
+                        (newDirector) =>
+                            !state.filteredDirectors.some((existingDirector) => existingDirector.id === newDirector.id)
                     );
-                    state.data = [...state.data, ...newData];
+                    state.filteredDirectors = [...state.filteredDirectors, ...newDirectors];
                 }
+
+                state.hasMore = action.payload.hasMore;
                 state.currentPage = action.payload.currentPage;
-                state.totalPages = action.payload.totalPages;
                 state.loading = false;
             })
-            .addCase(getContent.rejected, (state, action) => {
-                state.loading = false;
-                state.error = action.error.message;
-            })
-            .addCase(getFilteredContent.pending, (state) => {
-                state.loading = true;
-                state.error = null;
-            })
-            .addCase(getFilteredContent.fulfilled, (state, action) => {
-                if (action.payload.currentPage === 1) {
-                    state.data = action.payload.data;
-                } else {
-                    const newData = action.payload.data.filter(
-                        (newItem) => !state.data.some((existingItem) => existingItem.id === newItem.id)
-                    );
-                    state.data = [...state.data, ...newData];
-                }
-                state.currentPage = action.payload.currentPage;
-                state.totalPages = action.payload.totalPages;
-                state.loading = false;
-            })
-            .addCase(getFilteredContent.rejected, (state, action) => {
+            .addCase(getDirectorsByCountry.rejected, (state, action) => {
                 state.loading = false;
                 state.error = action.error.message;
             });
     },
 });
 
-export const movieActions = movieSlice.actions;
-export default movieSlice.reducer;
+export const { setSelectedCountry } = directorSlice.actions;
+export default directorSlice.reducer;
